Fix lazy scroll CSP and move hooks
